Fabric port: fix HyperVitalitySkill drawString, NeoForgeRegistries import, SynchedAnimationVariableKeys
- Fix HyperVitalitySkill drawString parenthesis
- Add NeoForgeRegistries import to EpicFightGlobalLootModifiers
- Fix EpicFightSynchedAnimationVariableKeys to use Keys.SYNCHED_ANIMATION_VARIABLE_KEY
- Down to ~196 total errors (syntax fix exposed hidden errors)

+++ src/main/java/yesman/epicfight/skill/passive/HyperVitalitySkill.java
@@ -104,7 +104,7 @@ public class HyperVitalitySkill extends PassiveSkill {
 
         if (!this.isActivated(container)) {
             String remainTime = String.format("%.0f", container.getMaxResource() - container.getResource());
-            guiGraphics.drawString(gui.getFont(), remainTime, (x + 12 - 4 * (remainTime.length())), y + 6, 16777215, true);
+            guiGraphics.drawString(gui.getFont(), remainTime, (int)(x + 12 - 4 * (remainTime.length())), (int)(y + 6), 16777215, true);
         }
     }
